Added return to intro state functionality to existing button.
The visibility of the intro dialog is now controlled by a variable in Generic.Variables called homepageVisible.
Because this variable is initialized to 1, the intro plays as normal the easyDiffraction is started.
When analysis is pressed, homepageVisible is set to 0, and the main program window becomes visible.
When the home button is pressed, homepageVisible is set to 1 and the intro view is shown.

The toolbarCurrentIndex was previous initialized to -1, and the analysis button set it to 0, corresponding to the project tab.
Now toolbarCurrentIndex is initialized to 0, and the analysis button does nothing. This means the tab is not changed when
going back and forth between the application window and the intro window.

This branch is based on cryspy_2.0.