Merged MeasuredDataModel and MeasuredDataSeries classes and updated the calls in Proxy.py to use the combined class instead.
Two bindings are still presented to QML, but both point to the same
instance of the MeasuredDataModel object.

It was found that the earlier version of MeasuredDataModel did not
support having multiple experiments, so this functionality was not
included in the new version either. This will need to be fixed in
a future update of this code.